common: fix conversion data warnings
Due to -Wconversion, a lot of conversion have been discovered.

+++ common/libs/VkCodecUtils/VulkanVideoUtils.cpp
@@ -793,11 +793,11 @@ VkResult VulkanPerDrawContext::RecordCommandBuffer(VkCommandBuffer cmdBuffer,
     vk::TransformPushConstants constants;
     if (scaleInput) {
         if (displayWidth && (displayWidth != inputImageToDrawFrom->imageWidth)) {
-            constants.texMatrix[0] = vk::Vec2((float)displayWidth / inputImageToDrawFrom->imageWidth, 0.0f);
+            constants.texMatrix[0] = vk::Vec2((float)(double)displayWidth / (float)(double)inputImageToDrawFrom->imageWidth, 0.0f);
         }
 
         if (displayHeight && (displayHeight != inputImageToDrawFrom->imageHeight)) {
-            constants.texMatrix[1] = vk::Vec2(.0f, (float)displayHeight /inputImageToDrawFrom->imageHeight);
+            constants.texMatrix[1] = vk::Vec2(.0f, (float)(double)displayHeight / (float)(double)inputImageToDrawFrom->imageHeight);
         }
     }
